feat(kubernetes-action): add support for the new backend system
