+++ plugins/kubernetes-actions/README.md
@@ -14,7 +14,7 @@ Run the following command to install the action package in your Backstage projec
 yarn workspace backend add @janus-idp/backstage-scaffolder-backend-module-kubernetes
 ```
 
-## Configuration
+### Installing the action on the legacy backend
 
 1. [Register](https://backstage.io/docs/features/software-templates/writing-custom-actions#registering-custom-actions) the Kubenretes actions by modifying the `packages/backend/src/plugins/scaffolder.ts` file from your project with the following changes:
 
@@ -71,19 +71,36 @@ yarn workspace backend add @janus-idp/backstage-scaffolder-backend-module-kubern
    yarn workspace backend add @backstage/integration
    ```
 
-3. Add the Kubernetes actions to your templates, see the [example](./examples/templates/01-kubernetes-template.yaml) file in this repository for complete usage examples
+### Installing the action on the new backend
 
-   ```yaml
-   action: kubernetes:create-namespace
-   id: create-kubernetes-namespace
-   name: Create kubernetes namespace
-   input:
-     namespace: foo
-     clusterRef: bar
-     token: TOKEN
-     skipTLSVerify: false
-     caData: Zm9v
-   ```
+Add the following to your `packages/backend/src/index.ts` file:
+
+```ts title="packages/backend/src/index.ts"
+const backend = createBackend();
+
+// Add the following line
+backend.add(
+  import('@janus-idp/backstage-scaffolder-backend-module-kubernetes/alpha'),
+);
+
+backend.start();
+```
+
+## Configuration
+
+Add the Kubernetes actions to your templates, see the [example](./examples/templates/01-kubernetes-template.yaml) file in this repository for complete usage examples
+
+```yaml
+action: kubernetes:create-namespace
+id: create-kubernetes-namespace
+name: Create kubernetes namespace
+input:
+  namespace: foo
+  clusterRef: bar
+  token: TOKEN
+  skipTLSVerify: false
+  caData: Zm9v
+```
 
 ## Usage
