+++ plugins/kubernetes-actions/dist-dynamic/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@janus-idp/backstage-scaffolder-backend-module-kubernetes-dynamic",
   "description": "The kubernetes module for @backstage/plugin-scaffolder-backend",
-  "version": "1.2.6",
+  "version": "1.2.7",
   "main": "dist/index.cjs.js",
   "types": "dist/index.d.ts",
   "license": "Apache-2.0",
@@ -12,6 +12,17 @@
   },
   "backstage": {
     "role": "backend-plugin-module"
+  },
+  "exports": {
+    ".": {
+      "require": "./dist/index.cjs.js",
+      "default": "./dist/index.cjs.js"
+    },
+    "./alpha": {
+      "require": "./dist/alpha.cjs.js",
+      "default": "./dist/alpha.cjs.js"
+    },
+    "./package.json": "./package.json"
   },
   "scripts": {
     "start": "backstage-cli package start",
@@ -29,7 +40,8 @@
   },
   "devDependencies": {},
   "files": [
-    "dist"
+    "dist",
+    "alpha"
   ],
   "repository": "github:janus-idp/backstage-plugins",
   "keywords": [
