+++ plugins/kubernetes-actions/src/alpha.ts
@@ -1,0 +1,1 @@
+export { scaffolderModuleKubernetesAction as default } from './module';
